requirements.txt
Django==1.4
South
Markdown
django-treebeard
python-twitter
simplejson
psycopg2
PIL

ipython
fabric